data-commit1.1
Refactored our state (for Vigor, right now) so that we stored a initState and the state that we will freely change. Triple confirm that this works, and apply this to all other pieces of state where we want to have a 'base' or init value.

PROBLEMS: apply ternary for the add and subtract function, work on making them reusable for all our pieces of state. apply the logic for the functions