data commit 1.3
Refactored our state (for Vigor, right now) so that we stored a initState and the state that we will freely change. Triple confirm that this works, and apply this to all other pieces of state where we want to have a 'base' or init value.

PROBLEMS: Our subtract function doesnt follow the logic once we incriment it .
Correct this.